% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupSession.R
\docType{package}
\name{backupSession}
\alias{backupSession}
\alias{backupSession-package}
\title{backupSession: Save and Load R Session Images, History and sessionInfo}
\description{
Save and load consistently named and versioned session images, history and sessionInfo().
This package will save and load three types of files: 1) R session images, 2) R command histories and 3) R sessionInfo() output.
See ?save.session, ?load.session and ?load.sinfo for usage examples and further information.
}
