% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupSession.R
\name{load.mdata}
\alias{load.mdata}
\title{load.mdata Function}
\usage{
load.mdata(basename = "NULL", version = "NULL", path = getwd(),
  verbose = FALSE)
}
\arguments{
\item{basename}{Basename for the R metadata file.}

\item{version}{A date string or version number used as part of the backup filename.  Cannot be an empty string.}

\item{path}{Directory to load file from.  Defaults to current working directory.}

\item{verbose}{Print file loading progress messages.  Must be either TRUE or FALSE.  Defaults to FALSE.}
}
\description{
This function loads R metadata (output of ls.str() function)
}
\examples{
\dontrun{
mData <- load.mdata(basename = "projectX", path = "./backups", version = "01.03.18.11.43")
}
}
\seealso{
\code{\link{load.session}}, \code{\link{save.session}} and \code{\link{load.sinfo}}
}
