% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupSession.R
\name{load.session}
\alias{load.session}
\title{load.session Function}
\usage{
load.session(basename = "NULL", version = "NULL", path = getwd(),
  verbose = FALSE)
}
\arguments{
\item{basename}{Basename for the R session images and history files.}

\item{version}{A date string or version number used as part of the backup filenames.  Cannot be an empty string.}

\item{path}{Directory to load backup files from.  Defaults to current working directory.}

\item{verbose}{Print session loading progress messages.  Must be either TRUE or FALSE.  Defaults to FALSE.}
}
\description{
This function loads R session images and if available history files saved with save.session().
}
\details{
It will overwrite objects in .GlobalEnv just as the load() function does.
History files are NOT loaded in non-interactive R sessions.
}
\examples{
\dontrun{
load.session(basename = "projectX", path = "./backups", version = "01.03.18.11.43")
}
}
\seealso{
\code{\link{save.session}}, \code{\link{load.mdata}} and \code{\link{load.sinfo}}
}
