% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupSession.R
\name{load.sinfo}
\alias{load.sinfo}
\title{load.sinfo Function}
\usage{
load.sinfo(basename = "NULL", version = "NULL", path = getwd(),
  verbose = FALSE)
}
\arguments{
\item{basename}{Basename for the R session info files.}

\item{version}{A date string or version number used as part of the backup filenames.  Cannot be an empty string.}

\item{path}{Directory to load files from.  Defaults to current working directory.}

\item{verbose}{Print file loading progress messages.  Must be either TRUE or FALSE.  Defaults to FALSE.}
}
\description{
This function loads R session information (output of sessionInfo() command)
}
\examples{
\dontrun{
sInfo <- load.sinfo(basename = "projectX", path = "./backups", version = "01.03.18.11.43")
}
}
\seealso{
[load.session()] and [save.session()]
}
