% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupSession.R
\docType{package}
\name{backupSession}
\alias{backupSession}
\alias{backupSession-package}
\title{backupSession: Save and Load R Session Images, History, metadata and sessionInfo}
\description{
Save and load consistently named and versioned session images, history, metadata and sessionInfo().
This package will save and load four types of files: 1) R session images, 2) R command histories, 3) R ls.str() output and 4) R sessionInfo() output.
See \code{\link{save.session}}, \code{\link{load.session}}, \code{\link{load.mdata}} and \code{\link{load.sinfo}} for usage examples and further information.
}
