% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupSession.R
\name{load.session}
\alias{load.session}
\title{load.session Function}
\usage{
load.session(basename = "NULL", path = "NULL", version = "NULL",
  verbose = FALSE)
}
\arguments{
\item{basename}{Basename for the R session images, history and session info files.}

\item{path}{Directory to load backup files from.  Defaults to current working directory.}

\item{version}{A date string or version number used as part of the backup filenames.  Cannot be an empty string.}

\item{verbose}{Print session loading progress messages.  Must be either TRUE or FALSE.  Defaults to FALSE.}
}
\description{
This function loads R session images, history and session info files saved with save.session().
}
\examples{
load.session(basename='projectX', path='./backups/', version='01.03.18.11.43')
}
