% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupSession.R
\name{save.session}
\alias{save.session}
\title{save.session Function}
\usage{
save.session(basename = "NULL", path = "NULL", version = "NULL",
  verbose = FALSE, force = FALSE)
}
\arguments{
\item{basename}{Basename for the R session images, history and session info files.}

\item{path}{Directory to save backup files to.  Defaults to current working directory.  Creates directories unless they exist.}

\item{version}{A date string or version number used as part of the backup filenames.  Cannot be an empty string.  Defaults to year.month.date.hour.minute formatted timestamps.}

\item{verbose}{Print session loading progress messages.  Must be either TRUE or FALSE.  Defaults to FALSE.}

\item{force}{Overwrite existing files.  Must be either TRUE or FALSE.  Defaults to FALSE.}
}
\description{
This function saves R session images, history and session info files which can be loaded with load.session().
}
\examples{
save.session(basename='projectX', path='./backups/', version='01.03.18.11.43')
}
